Added minecraft support back into the calculator
minecraft's resources.yaml was auto-generated using a json to yaml tool with the
source json being taken from the old ruby site's generated json. This has lead
to the file being not as clean as the astroneer file and will need to be
cleaned up before releasing it to the public in order to keep standards.