Completed new dependency tracking
The new system can properly track dependencies and does not rely on an input file to fully dictate what the output files are. Instead all input and output files are defined dynamically based on a regex match to some subset of the input files. Of which the existance of those files would indicate that an output file should also exist