Propagating the schema changes into the codebase
Two calculators have diffs, minecraft and stardew valley. Both of these diffs were investigated manually and in both cases it was due to a Raw Resource recipe being present but not being first or last in the recipe list. With the new boolean selector for raw recipes a raw recipe will only ever be first or last in the list. In all cases the only difference was the raw recipe moving from its original position to the end of the list of recipes